% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{pd_latest}
\alias{pd_latest}
\title{The newest version of a dataset}
\usage{
pd_latest(slug)
}
\arguments{
\item{slug}{The dataset's slug, as in its page URL `https://publicdata.au/d/<slug>/`.}
}
\value{
The newest version's date as text, such as `"2026-09-28"`.
}
\description{
The newest version of a dataset
}
\examples{
\dontshow{if (pd_available()) withAutoprint(\{ # examplesIf}
pd_latest("au-road-deaths")
\dontshow{\}) # examplesIf}
}
\seealso{
Other metadata:
\code{\link[=pd_attribution]{pd_attribution()}},
\code{\link[=pd_changes]{pd_changes()}},
\code{\link[=pd_cite]{pd_cite()}},
\code{\link[=pd_fields]{pd_fields()}},
\code{\link[=pd_provenance]{pd_provenance()}},
\code{\link[=pd_schema]{pd_schema()}},
\code{\link[=pd_tables]{pd_tables()}}
}
\concept{metadata}
